Add files via upload
In using the delimiter, I was able to successfully separate "category" and "sub-category", however, somehow my "sub-category" labels do not reflect the effect of the delimiter that you can see in the "RawData" tab. My apologies for this, but I could not find a way to solve for this. I hope this still demonstrates what is being requested through the homework. thank you!